+++ src/index.tsx
@@ -1,10 +1,14 @@
 export {
-  fetchConfig,
-  fetchConfigRequest,
-  fetchConfigSuccess,
-  fetchConfigFailure,
-  fetchConfigCloseErrorModalRequest
+    fetchConfig,
+    fetchConfigRequest,
+    fetchConfigSuccess,
+    fetchConfigFailure,
 }from './js/fetch-config/actions-creators';
 export {fetchConfigReducer} from './js/fetch-config/reducers';
 export {FetchConfigState} from './js/fetch-config/types';
-export {WsBrand} from './js/components/wsbrand/WsBrand';
+export {WsBrand} from './js/components/ws-brand/WsBrand';
+export {
+    unexpectedErrorOccured,
+}from './js/components/ws-unexpected-error/actions-creators';
+export {unexpectedErrorReducer} from './js/components/ws-unexpected-error/reducers';
+export {WsUnexpectedError} from './js/components/ws-unexpected-error/WsUnexpectedError';