+++ src/js/fetch-config/actions-creators.tsx
@@ -1,4 +1,5 @@
 import {FETCH_CONFIG} from './types';
+import {unexpectedErrorOccured} from '../components/ws-unexpected-error/actions-creators';
 
 const fetchConfig = () => (dispatch: Function) => {
     dispatch(fetchConfigRequest());
@@ -8,12 +9,14 @@ const fetchConfig = () => (dispatch: Function) => {
                 return response.json().then(data => dispatch(fetchConfigSuccess(data)));
             } else {
                 dispatch(fetchConfigFailure());
+                dispatch(unexpectedErrorOccured());
                 return;
             }
         })
         .catch(e => {
             console.error(e);
             dispatch(fetchConfigFailure());
+            dispatch(unexpectedErrorOccured());
         })
 };
 
@@ -23,12 +26,9 @@ const fetchConfigSuccess = <T extends {}>(data: T) => ({type: FETCH_CONFIG.SUCCE
 
 const fetchConfigFailure = () => ({type: FETCH_CONFIG.FAILURE});
 
-const fetchConfigCloseErrorModalRequest = () => ({type: FETCH_CONFIG.CLOSE_ERROR_MODAL_REQUEST});
-
 export {
     fetchConfig,
     fetchConfigRequest,
     fetchConfigSuccess,
     fetchConfigFailure,
-    fetchConfigCloseErrorModalRequest
 }