+++ src/js/fetch-config/reducers.tsx
@@ -3,8 +3,6 @@ import {FetchConfigAction, FetchConfigState, FETCH_CONFIG} from './types';
 const initialState = {
   isLoading: false,
   data: null,
-  failed: false,
-  showModal: false,
 };
 
 export const fetchConfigReducer = <T extends {}>(
@@ -12,13 +10,11 @@ export const fetchConfigReducer = <T extends {}>(
   action: FetchConfigAction<T>): FetchConfigState<T> => {
   switch (action.type) {
     case FETCH_CONFIG.REQUEST:
-      return {...state, isLoading: true, data: null, failed: false, showModal: false};
+      return {...state, isLoading: true, data: null};
     case FETCH_CONFIG.SUCCESS:
-      return {...state, isLoading: false, data: action.data, failed: false, showModal: false};
+      return {...state, isLoading: false, data: action.data};
     case FETCH_CONFIG.FAILURE:
-      return {...state, isLoading: false, data: null, failed: true, showModal: true};
-    case FETCH_CONFIG.CLOSE_ERROR_MODAL_REQUEST:
-      return {...state, isLoading: false, data: null, failed: true, showModal: false};
+      return {...state, isLoading: false, data: null};
     default:
       return state;
   }