+++ src/test/components/wsbrand/WsBrand.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {WsBrand} from '../../../js/components/wsbrand/WsBrand';
+import {WsBrand} from '../../../js/components/ws-brand/WsBrand';
 import {createRenderer} from 'react-test-renderer/shallow';
 
 it('should render', () => {