+++ CHANGELOG.md
@@ -2,9 +2,23 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.2] - 2019-03-21
+### Fixed
+- Exports
+
+## [3.0.1] - 2019-03-21
+### Fixed
+- Don't close unexpected error when closable is set
+
+## [3.0.0] - 2019-03-21
+### Added
+- WsUnexpectedError
+### Removed
+- showModal from fetch config state
+
 ## [2.0.1] - 2019-03-20
 ### Fixed
-Published version
+- Published version
 
 ## [2.0.0] - 2019-03-20
 ### Removed